requirements.txt
fastmcp
keycardai[mcp-fastmcp]
crewai[tools]
httpx
uvicorn[standard]
python-dotenv
